refactor(Profile) : Modal Header의 세그먼트로 이동
