--- src/1_Page/Profile/model/useModifySettingClick.js
@@ -1,16 +1,0 @@
-import { useState } from "react";
-
-const useModifySettingClick = () => {
-  const [isModifyClick, setIsmodifyClick] = useState(false);
-
-  const handleModifyClickFalse = () => setIsmodifyClick(false);
-  const handleModalModifyTrue = () => setIsmodifyClick(true);
-
-  return {
-    isModifyClick,
-    handleModifyClickFalse,
-    handleModalModifyTrue,
-  };
-};
-
-export default useModifySettingClick;
