+++ src/1_Page/Profile/ui/Header/model/useModifyMode.js
@@ -1,0 +1,12 @@
+import { useState } from "react";
+
+const useModifyMode = () => {
+  const [modifyModeModal, setModifyModeModal] = useState(false);
+
+  const handleModifyModeClose = () => setModifyModeModal(false);
+  const handleModifyModeOpen = () => setModifyModeModal(true);
+
+  return { modifyModeModal, handleModifyModeClose, handleModifyModeOpen };
+};
+
+export default useModifyMode;
